binary_deployment/requirements.txt
torch>=1.11.0
torchvision>=0.12.0
Pillow>=8.0.0